+++ src/Streamly/Internal/Data/Stream/Ahead.hs
@@ -786,7 +786,7 @@ apAhead (AheadT m1) (AheadT m2) =
 
 instance (Monad m, MonadAsync m) => Applicative (AheadT m) where
     {-# INLINE pure #-}
-    pure = AheadT . K.yield
+    pure = AheadT . K.fromPure
     {-# INLINE (<*>) #-}
     (<*>) = apAhead